requirements.txt
groq>=0.12.0
python-dotenv>=1.0.0
gradio>=4.44.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0